Restore proof serialization helpers
