+++ src/proof/ser.rs
@@ -29,18 +29,15 @@
 //! data. All helpers operate exclusively through these wrapper APIs so callers
 //! preserve the integrity checks enforced by the handles.
 
-use crate::config::{AirSpecId, ParamDigest, ProofKind};
-use crate::fri::FriProof;
+use crate::config::{ParamDigest, ProofKind};
 use crate::hash::Hasher;
 use crate::proof::public_inputs::{
     AggregationHeaderV1, ExecutionHeaderV1, ProofKind as PublicProofKind, PublicInputVersion,
     PublicInputs, RecursionHeaderV1, VrfHeaderV1,
 };
 use crate::proof::types::{
-    CompositionBinding, CompositionOpenings, FriHandle, FriParametersMirror,
-    MerkleAuthenticationPath, MerklePathNode, MerkleProofBundle, MerkleSection, Openings,
-    OpeningsDescriptor, OutOfDomainOpening, Proof, Telemetry, TelemetryOption, TraceOpenings,
-    VerifyError, PROOF_VERSION,
+    CompositionBinding, FriHandle, MerkleSection, Openings, OpeningsDescriptor, OutOfDomainOpening,
+    Proof, TelemetryOption, VerifyError, PROOF_VERSION,
 };
 use crate::ser::{
     ensure_consumed, ensure_u32, read_digest, read_u16, read_u32, read_u8, write_bytes,
@@ -183,31 +180,6 @@ pub(crate) fn encode_proof_kind(kind: ProofKind) -> u8 {
         ProofKind::VRF => 7,
     }
 }
-
-pub(crate) fn decode_proof_kind(byte: u8) -> Result<ProofKind, VerifyError> {
-    Ok(match byte {
-        0 => ProofKind::Tx,
-        1 => ProofKind::State,
-        2 => ProofKind::Pruning,
-        3 => ProofKind::Uptime,
-        4 => ProofKind::Consensus,
-        5 => ProofKind::Identity,
-        6 => ProofKind::Aggregation,
-        7 => ProofKind::VRF,
-        other => return Err(VerifyError::UnknownProofKind(other)),
-    })
-}
-
-pub(crate) fn serialize_fri_proof(proof: &FriProof) -> Result<Vec<u8>, SerError> {
-    proof
-        .to_bytes()
-        .map_err(|_| SerError::invalid_value(SerKind::Fri, "fri_proof"))
-}
-
-pub(crate) fn deserialize_fri_proof(bytes: &[u8]) -> Result<FriProof, VerifyError> {
-    FriProof::from_bytes(bytes).map_err(|_| VerifyError::Serialization(SerKind::Fri))
-}
-
 pub fn compute_public_digest(bytes: &[u8]) -> [u8; DIGEST_SIZE] {
     let mut hasher = Hasher::new();
     hasher.update(bytes);
@@ -313,10 +285,10 @@ pub fn serialize_proof(proof: &Proof) -> Result<Vec<u8>, SerError> {
         return Err(map_composition_issue_to_ser(issue));
     }
 
-    let binding_bytes = serialize_composition_binding(binding)?;
-    let openings_bytes = serialize_openings_descriptor(openings_descriptor)?;
-    let fri_bytes = serialize_fri_handle(proof.fri())?;
-    let telemetry_bytes = serialize_telemetry_option(proof.telemetry())?;
+    let binding_bytes = binding.serialize_bytes()?;
+    let openings_bytes = openings_descriptor.serialize_bytes()?;
+    let fri_bytes = proof.fri().serialize_bytes()?;
+    let telemetry_bytes = proof.telemetry().serialize_bytes()?;
 
     let header = serialize_proof_header_from_lengths(
         proof,
@@ -339,6 +311,49 @@ pub fn serialize_proof(proof: &Proof) -> Result<Vec<u8>, SerError> {
     }
 
     Ok(out)
+}
+
+/// Serialises the proof header given the payload bytes.
+pub fn serialize_proof_header(proof: &Proof, payload: &[u8]) -> Result<Vec<u8>, SerError> {
+    let binding_bytes = proof.composition().serialize_bytes()?;
+    let openings_bytes = proof.openings().serialize_bytes()?;
+    let fri_bytes = proof.fri().serialize_bytes()?;
+    let telemetry_bytes = proof.telemetry().serialize_bytes()?;
+
+    let telemetry_total = telemetry_bytes.as_ref().map_or(0, |bytes| bytes.len());
+    let expected_total = openings_bytes.len() + fri_bytes.len() + telemetry_total;
+    if payload.len() != expected_total {
+        return Err(SerError::invalid_value(
+            SerKind::Proof,
+            "payload_length_mismatch",
+        ));
+    }
+
+    serialize_proof_header_from_lengths(
+        proof,
+        &binding_bytes,
+        openings_bytes.len(),
+        fri_bytes.len(),
+        telemetry_bytes.as_ref().map(|bytes| bytes.len()),
+    )
+}
+
+/// Serialises the proof payload (body) without the integrity digest.
+pub fn serialize_proof_payload(proof: &Proof) -> Result<Vec<u8>, SerError> {
+    let openings_bytes = proof.openings().serialize_bytes()?;
+    let fri_bytes = proof.fri().serialize_bytes()?;
+    let telemetry_bytes = proof.telemetry().serialize_bytes()?;
+
+    let capacity = openings_bytes.len()
+        + fri_bytes.len()
+        + telemetry_bytes.as_ref().map_or(0, |bytes| bytes.len());
+    let mut buffer = Vec::with_capacity(capacity);
+    buffer.extend_from_slice(&openings_bytes);
+    buffer.extend_from_slice(&fri_bytes);
+    if let Some(bytes) = telemetry_bytes {
+        buffer.extend_from_slice(&bytes);
+    }
+    Ok(buffer)
 }
 
 fn serialize_proof_header_from_lengths(
@@ -407,7 +422,7 @@ pub fn deserialize_proof(bytes: &[u8]) -> Result<Proof, VerifyError> {
     let binding_bytes = cursor
         .read_vec(SerKind::Proof, "composition_bytes", binding_len)
         .map_err(VerifyError::from)?;
-    let binding = deserialize_composition_binding(&binding_bytes)?;
+    let binding = CompositionBinding::deserialize_bytes(&binding_bytes)?;
 
     let openings_len = read_u32(&mut cursor, SerKind::Openings, "descriptor_len")
         .map_err(VerifyError::from)? as usize;
@@ -429,21 +444,20 @@ pub fn deserialize_proof(bytes: &[u8]) -> Result<Proof, VerifyError> {
     let openings_bytes = cursor
         .read_vec(SerKind::Openings, "descriptor_bytes", openings_len)
         .map_err(VerifyError::from)?;
-    let openings_descriptor =
-        deserialize_openings_descriptor(&openings_bytes).map_err(VerifyError::from)?;
+    let openings_descriptor = OpeningsDescriptor::deserialize_bytes(&openings_bytes)?;
 
     let fri_bytes = cursor
         .read_vec(SerKind::Fri, "fri_bytes", fri_len)
         .map_err(VerifyError::from)?;
-    let fri_handle = deserialize_fri_handle(&fri_bytes)?;
+    let fri_handle = FriHandle::deserialize_bytes(&fri_bytes)?;
 
     let telemetry_option = if let Some(len) = telemetry_len {
         let telemetry_bytes = cursor
             .read_vec(SerKind::Telemetry, "telemetry_bytes", len)
             .map_err(VerifyError::from)?;
-        deserialize_telemetry_option(telemetry_present, Some(&telemetry_bytes))?
+        TelemetryOption::deserialize_bytes(telemetry_present, Some(&telemetry_bytes))?
     } else {
-        deserialize_telemetry_option(false, None)?
+        TelemetryOption::deserialize_bytes(false, None)?
     };
 
     ensure_consumed(&cursor, SerKind::Proof).map_err(VerifyError::from)?;
@@ -478,429 +492,19 @@ pub fn deserialize_proof(bytes: &[u8]) -> Result<Proof, VerifyError> {
     ))
 }
 
-fn serialize_composition_binding(binding: &CompositionBinding) -> Result<Vec<u8>, SerError> {
-    let mut buffer = Vec::new();
-    write_u8(&mut buffer, encode_proof_kind(*binding.kind()));
-    write_digest(&mut buffer, binding.air_spec_id().as_bytes());
-
-    let public_inputs = binding.public_inputs();
-    let public_len = ensure_u32(public_inputs.len(), SerKind::PublicInputs, "len")?;
-    write_u32(&mut buffer, public_len);
-    write_bytes(&mut buffer, public_inputs);
-
-    match binding.composition_commit() {
-        Some(commit) => {
-            write_u8(&mut buffer, 1);
-            write_digest(&mut buffer, &commit.bytes);
-        }
-        None => write_u8(&mut buffer, 0),
-    }
-
-    Ok(buffer)
-}
-
-fn deserialize_composition_binding(bytes: &[u8]) -> Result<CompositionBinding, VerifyError> {
-    let mut cursor = ByteReader::new(bytes);
-    let kind_byte = read_u8(&mut cursor, SerKind::Proof, "kind").map_err(VerifyError::from)?;
-    let kind = decode_proof_kind(kind_byte)?;
-
-    let air_spec_id = AirSpecId(DigestBytes {
-        bytes: read_digest(&mut cursor, SerKind::Proof, "air_spec_id")
-            .map_err(VerifyError::from)?,
-    });
-
-    let public_len =
-        read_u32(&mut cursor, SerKind::PublicInputs, "len").map_err(VerifyError::from)? as usize;
-    let public_inputs = cursor
-        .read_vec(SerKind::PublicInputs, "public_inputs", public_len)
-        .map_err(VerifyError::from)?;
-
-    let composition_commit = match read_u8(&mut cursor, SerKind::CompositionCommitment, "flag")
-        .map_err(VerifyError::from)?
-    {
-        0 => None,
-        1 => Some(DigestBytes {
-            bytes: read_digest(&mut cursor, SerKind::CompositionCommitment, "digest")
-                .map_err(VerifyError::from)?,
-        }),
-        _ => return Err(VerifyError::Serialization(SerKind::CompositionCommitment)),
-    };
-
-    ensure_consumed(&cursor, SerKind::Proof).map_err(VerifyError::from)?;
-
-    Ok(CompositionBinding::new(
-        kind,
-        air_spec_id,
-        public_inputs,
-        composition_commit,
-    ))
-}
-
-fn serialize_openings_descriptor(descriptor: &OpeningsDescriptor) -> Result<Vec<u8>, SerError> {
-    let merkle_bytes = serialize_merkle_bundle(descriptor.merkle())?;
-    let openings_bytes = serialize_openings(descriptor.openings())?;
-
-    let mut buffer = Vec::new();
-    let merkle_len = ensure_u32(merkle_bytes.len(), SerKind::TraceCommitment, "merkle_len")?;
-    write_u32(&mut buffer, merkle_len);
-    write_bytes(&mut buffer, &merkle_bytes);
-
-    let openings_len = ensure_u32(openings_bytes.len(), SerKind::Openings, "openings_len")?;
-    write_u32(&mut buffer, openings_len);
-    write_bytes(&mut buffer, &openings_bytes);
-
-    Ok(buffer)
-}
-
-fn deserialize_openings_descriptor(bytes: &[u8]) -> Result<OpeningsDescriptor, SerError> {
-    let mut cursor = ByteReader::new(bytes);
-    let merkle_len = read_u32(&mut cursor, SerKind::TraceCommitment, "merkle_len")? as usize;
-    let merkle_bytes = cursor.read_vec(SerKind::TraceCommitment, "merkle_bytes", merkle_len)?;
-
-    let openings_len = read_u32(&mut cursor, SerKind::Openings, "openings_len")? as usize;
-    let openings_bytes = cursor.read_vec(SerKind::Openings, "openings_bytes", openings_len)?;
-
-    ensure_consumed(&cursor, SerKind::Proof)?;
-
-    let merkle = deserialize_merkle_bundle(&merkle_bytes)?;
-    let openings = deserialize_openings(&openings_bytes)?;
-    Ok(OpeningsDescriptor::new(merkle, openings))
-}
-
-fn serialize_fri_handle(handle: &FriHandle) -> Result<Vec<u8>, SerError> {
-    serialize_fri_proof(handle.fri_proof())
-}
-
-fn deserialize_fri_handle(bytes: &[u8]) -> Result<FriHandle, VerifyError> {
-    deserialize_fri_proof(bytes).map(FriHandle::new)
-}
-
-fn serialize_telemetry_option(option: &TelemetryOption) -> Result<Option<Vec<u8>>, SerError> {
-    if option.is_present() {
-        Ok(Some(serialize_telemetry_frame(option.frame())?))
-    } else {
-        Ok(None)
-    }
-}
-
-fn deserialize_telemetry_option(
-    present: bool,
-    bytes: Option<&[u8]>,
-) -> Result<TelemetryOption, VerifyError> {
-    if present {
-        let telemetry_bytes =
-            bytes.ok_or_else(|| VerifyError::Serialization(SerKind::Telemetry))?;
-        let telemetry = deserialize_telemetry_frame(telemetry_bytes).map_err(VerifyError::from)?;
-        Ok(TelemetryOption::new(true, telemetry))
-    } else {
-        Ok(TelemetryOption::new(false, Telemetry::default()))
-    }
-}
-
-fn serialize_merkle_bundle(bundle: &MerkleProofBundle) -> Result<Vec<u8>, SerError> {
-    let mut out = Vec::new();
-    write_digest(&mut out, bundle.core_root());
-    write_digest(&mut out, bundle.aux_root());
-    let layer_count = ensure_u32(
-        bundle.fri_layer_roots().len(),
-        SerKind::TraceCommitment,
-        "fri_roots",
-    )?;
-    write_u32(&mut out, layer_count);
-    for root in bundle.fri_layer_roots() {
-        write_digest(&mut out, root);
-    }
-    Ok(out)
-}
-
-fn deserialize_merkle_bundle(bytes: &[u8]) -> Result<MerkleProofBundle, SerError> {
-    let mut cursor = ByteReader::new(bytes);
-    let core_root = read_digest(&mut cursor, SerKind::TraceCommitment, "core_root")?;
-    let aux_root = read_digest(&mut cursor, SerKind::TraceCommitment, "aux_root")?;
-    let layer_count = read_u32(&mut cursor, SerKind::TraceCommitment, "fri_roots")? as usize;
-    let mut fri_layer_roots = Vec::with_capacity(layer_count);
-    for _ in 0..layer_count {
-        fri_layer_roots.push(read_digest(
-            &mut cursor,
-            SerKind::TraceCommitment,
-            "fri_root",
-        )?);
-    }
-    ensure_consumed(&cursor, SerKind::TraceCommitment)?;
-    Ok(MerkleProofBundle::new(core_root, aux_root, fri_layer_roots))
-}
-
-fn encode_openings(openings: &Openings) -> Result<Vec<u8>, SerError> {
-    let mut buffer = Vec::new();
-    encode_merkle_openings(
-        &mut buffer,
-        openings.trace().indices(),
-        openings.trace().leaves(),
-        openings.trace().paths(),
-    )?;
-    match openings.composition() {
-        Some(section) => {
-            write_u8(&mut buffer, 1);
-            encode_merkle_openings(
-                &mut buffer,
-                section.indices(),
-                section.leaves(),
-                section.paths(),
-            )?;
-        }
-        None => write_u8(&mut buffer, 0),
-    }
-
-    let count = ensure_u32(openings.out_of_domain().len(), SerKind::Openings, "ood_len")?;
-    write_u32(&mut buffer, count);
-    for opening in openings.out_of_domain() {
-        let encoded = serialize_out_of_domain_opening(opening)?;
-        let encoded_len = ensure_u32(encoded.len(), SerKind::Openings, "ood_block")?;
-        write_u32(&mut buffer, encoded_len);
-        write_bytes(&mut buffer, &encoded);
-    }
-    Ok(buffer)
-}
-
-fn deserialize_openings(bytes: &[u8]) -> Result<Openings, SerError> {
-    let mut cursor = ByteReader::new(bytes);
-    let (trace_indices, trace_leaves, trace_paths) = decode_merkle_openings(&mut cursor)?;
-    let trace = TraceOpenings {
-        indices: trace_indices,
-        leaves: trace_leaves,
-        paths: trace_paths,
-    };
-    let has_composition = read_u8(&mut cursor, SerKind::Openings, "composition_flag")?;
-    let composition = match has_composition {
-        0 => None,
-        1 => {
-            let (indices, leaves, paths) = decode_merkle_openings(&mut cursor)?;
-            Some(CompositionOpenings {
-                indices,
-                leaves,
-                paths,
-            })
-        }
-        _ => {
-            return Err(SerError::invalid_value(
-                SerKind::Openings,
-                "composition_flag",
-            ))
-        }
-    };
-
-    let count = read_u32(&mut cursor, SerKind::Openings, "ood_len")? as usize;
-    let mut out = Vec::with_capacity(count);
-    for _ in 0..count {
-        let block_len = read_u32(&mut cursor, SerKind::Openings, "ood_block")? as usize;
-        let block = cursor.read_vec(SerKind::Openings, "ood_bytes", block_len)?;
-        let opening = deserialize_out_of_domain_opening_inner(&block)?;
-        out.push(opening);
-    }
-    ensure_consumed(&cursor, SerKind::Openings)?;
-    Ok(Openings {
-        trace,
-        composition,
-        out_of_domain: out,
-    })
-}
-
-fn serialize_telemetry_frame(telemetry: &Telemetry) -> Result<Vec<u8>, SerError> {
-    let mut out = Vec::new();
-    write_u32(&mut out, telemetry.header_length());
-    write_u32(&mut out, telemetry.body_length());
-    write_u8(&mut out, telemetry.fri_parameters().fold);
-    write_u16(&mut out, telemetry.fri_parameters().cap_degree);
-    write_u32(&mut out, telemetry.fri_parameters().cap_size);
-    write_u16(&mut out, telemetry.fri_parameters().query_budget);
-    write_digest(&mut out, &telemetry.integrity_digest().bytes);
-    Ok(out)
-}
-
-fn deserialize_telemetry_frame(bytes: &[u8]) -> Result<Telemetry, SerError> {
-    let mut cursor = ByteReader::new(bytes);
-    let header_length = read_u32(&mut cursor, SerKind::Telemetry, "header_length")?;
-    let body_length = read_u32(&mut cursor, SerKind::Telemetry, "body_length")?;
-    let fold = read_u8(&mut cursor, SerKind::Telemetry, "fri.fold")?;
-    let cap_degree = read_u16(&mut cursor, SerKind::Telemetry, "fri.cap_degree")?;
-    let cap_size = read_u32(&mut cursor, SerKind::Telemetry, "fri.cap_size")?;
-    let query_budget = read_u16(&mut cursor, SerKind::Telemetry, "fri.query_budget")?;
-    let integrity_digest = read_digest(&mut cursor, SerKind::Telemetry, "integrity_digest")?;
-    ensure_consumed(&cursor, SerKind::Telemetry)?;
-    Ok(Telemetry {
-        header_length,
-        body_length,
-        fri_parameters: FriParametersMirror {
-            fold,
-            cap_degree,
-            cap_size,
-            query_budget,
-        },
-        integrity_digest: DigestBytes {
-            bytes: integrity_digest,
-        },
-    })
-}
-
 /// Serialises the out-of-domain opening container.
 pub fn serialize_openings(openings: &Openings) -> Result<Vec<u8>, SerError> {
-    encode_openings(openings)
+    openings.serialize_bytes()
 }
 
 /// Serialises a single out-of-domain opening block.
 pub fn serialize_out_of_domain_opening(opening: &OutOfDomainOpening) -> Result<Vec<u8>, SerError> {
-    let mut buffer = Vec::new();
-    write_digest(&mut buffer, &opening.point);
-    let core_len = ensure_u32(opening.core_values.len(), SerKind::Openings, "core_len")?;
-    write_u32(&mut buffer, core_len);
-    for value in &opening.core_values {
-        write_digest(&mut buffer, value);
-    }
-    let aux_len = ensure_u32(opening.aux_values.len(), SerKind::Openings, "aux_len")?;
-    write_u32(&mut buffer, aux_len);
-    for value in &opening.aux_values {
-        write_digest(&mut buffer, value);
-    }
-    write_digest(&mut buffer, &opening.composition_value);
-    Ok(buffer)
-}
-
-fn deserialize_out_of_domain_opening_inner(bytes: &[u8]) -> Result<OutOfDomainOpening, SerError> {
-    let mut cursor = ByteReader::new(bytes);
-    let point = read_digest(&mut cursor, SerKind::Openings, "point")?;
-    let core_len = read_u32(&mut cursor, SerKind::Openings, "core_len")? as usize;
-    let mut core_values = Vec::with_capacity(core_len);
-    for _ in 0..core_len {
-        core_values.push(read_digest(&mut cursor, SerKind::Openings, "core_value")?);
-    }
-    let aux_len = read_u32(&mut cursor, SerKind::Openings, "aux_len")? as usize;
-    let mut aux_values = Vec::with_capacity(aux_len);
-    for _ in 0..aux_len {
-        aux_values.push(read_digest(&mut cursor, SerKind::Openings, "aux_value")?);
-    }
-    let composition_value = read_digest(&mut cursor, SerKind::Openings, "composition_value")?;
-    ensure_consumed(&cursor, SerKind::Openings)?;
-    Ok(OutOfDomainOpening {
-        point,
-        core_values,
-        aux_values,
-        composition_value,
-    })
-}
-
-fn encode_merkle_openings(
-    buffer: &mut Vec<u8>,
-    indices: &[u32],
-    leaves: &[Vec<u8>],
-    paths: &[MerkleAuthenticationPath],
-) -> Result<(), SerError> {
-    let indices_len = ensure_u32(indices.len(), SerKind::Openings, "indices_len")?;
-    write_u32(buffer, indices_len);
-    for index in indices {
-        write_u32(buffer, *index);
-    }
-
-    let leaves_len = ensure_u32(leaves.len(), SerKind::Openings, "leaves_len")?;
-    write_u32(buffer, leaves_len);
-    for leaf in leaves {
-        let leaf_len = ensure_u32(leaf.len(), SerKind::Openings, "leaf_len")?;
-        write_u32(buffer, leaf_len);
-        write_bytes(buffer, leaf);
-    }
-
-    let paths_len = ensure_u32(paths.len(), SerKind::Openings, "paths_len")?;
-    write_u32(buffer, paths_len);
-    for path in paths {
-        let nodes_len = ensure_u32(path.nodes.len(), SerKind::Openings, "path_len")?;
-        write_u32(buffer, nodes_len);
-        for node in &path.nodes {
-            let MerklePathNode { index, sibling } = node;
-            write_u8(buffer, *index);
-            write_digest(buffer, sibling);
-        }
-    }
-    Ok(())
-}
-
-type DecodedMerkleOpenings = (Vec<u32>, Vec<Vec<u8>>, Vec<MerkleAuthenticationPath>);
-
-fn decode_merkle_openings(cursor: &mut ByteReader<'_>) -> Result<DecodedMerkleOpenings, SerError> {
-    let indices_len = read_u32(cursor, SerKind::Openings, "indices_len")? as usize;
-    let mut indices = Vec::with_capacity(indices_len);
-    for _ in 0..indices_len {
-        indices.push(read_u32(cursor, SerKind::Openings, "index")?);
-    }
-
-    let leaves_len = read_u32(cursor, SerKind::Openings, "leaves_len")? as usize;
-    let mut leaves = Vec::with_capacity(leaves_len);
-    for _ in 0..leaves_len {
-        let leaf_len = read_u32(cursor, SerKind::Openings, "leaf_len")? as usize;
-        let bytes = cursor.read_vec(SerKind::Openings, "leaf_bytes", leaf_len)?;
-        leaves.push(bytes);
-    }
-
-    let paths_len = read_u32(cursor, SerKind::Openings, "paths_len")? as usize;
-    let mut paths = Vec::with_capacity(paths_len);
-    for _ in 0..paths_len {
-        let nodes_len = read_u32(cursor, SerKind::Openings, "path_len")? as usize;
-        let mut nodes = Vec::with_capacity(nodes_len);
-        for _ in 0..nodes_len {
-            let index = read_u8(cursor, SerKind::Openings, "path_index")?;
-            let sibling = read_digest(cursor, SerKind::Openings, "path_sibling")?;
-            nodes.push(MerklePathNode { index, sibling });
-        }
-        paths.push(MerkleAuthenticationPath { nodes });
-    }
-
-    Ok((indices, leaves, paths))
+    opening.serialize_bytes()
 }
 
 /// Deserialises an out-of-domain opening block.
 pub fn deserialize_out_of_domain_opening(bytes: &[u8]) -> Result<OutOfDomainOpening, VerifyError> {
-    deserialize_out_of_domain_opening_inner(bytes).map_err(VerifyError::from)
-}
-
-/// Serialises the proof header given the payload bytes.
-pub fn serialize_proof_header(proof: &Proof, payload: &[u8]) -> Result<Vec<u8>, SerError> {
-    let binding_bytes = serialize_composition_binding(proof.composition())?;
-    let openings_bytes = serialize_openings_descriptor(proof.openings())?;
-    let fri_bytes = serialize_fri_handle(proof.fri())?;
-    let telemetry_bytes = serialize_telemetry_option(proof.telemetry())?;
-
-    let telemetry_total = telemetry_bytes.as_ref().map_or(0, |bytes| bytes.len());
-    let expected_total = openings_bytes.len() + fri_bytes.len() + telemetry_total;
-    if payload.len() != expected_total {
-        return Err(SerError::invalid_value(
-            SerKind::Proof,
-            "payload_length_mismatch",
-        ));
-    }
-
-    serialize_proof_header_from_lengths(
-        proof,
-        &binding_bytes,
-        openings_bytes.len(),
-        fri_bytes.len(),
-        telemetry_bytes.as_ref().map(|bytes| bytes.len()),
-    )
-}
-
-/// Serialises the proof payload (body) without the integrity digest.
-pub fn serialize_proof_payload(proof: &Proof) -> Result<Vec<u8>, SerError> {
-    let openings_bytes = serialize_openings_descriptor(proof.openings())?;
-    let fri_bytes = serialize_fri_handle(proof.fri())?;
-    let telemetry_bytes = serialize_telemetry_option(proof.telemetry())?;
-
-    let capacity = openings_bytes.len()
-        + fri_bytes.len()
-        + telemetry_bytes.as_ref().map_or(0, |bytes| bytes.len());
-    let mut buffer = Vec::with_capacity(capacity);
-    buffer.extend_from_slice(&openings_bytes);
-    buffer.extend_from_slice(&fri_bytes);
-    if let Some(bytes) = telemetry_bytes {
-        buffer.extend_from_slice(&bytes);
-    }
-    Ok(buffer)
+    OutOfDomainOpening::deserialize_bytes(bytes)
 }
 
 #[cfg(test)]
@@ -912,8 +516,9 @@ mod tests {
     use crate::proof::params::canonical_stark_params;
     use crate::proof::public_inputs::{ExecutionHeaderV1, PublicInputVersion, PublicInputs};
     use crate::proof::types::{
-        CompositionBinding, CompositionOpenings, FriHandle, MerkleAuthenticationPath,
-        MerklePathNode, OpeningsDescriptor, TelemetryOption, TraceOpenings,
+        CompositionBinding, CompositionOpenings, FriHandle, FriParametersMirror,
+        MerkleAuthenticationPath, MerklePathNode, MerkleProofBundle, OpeningsDescriptor, Telemetry,
+        TelemetryOption, TraceOpenings,
     };
     use crate::utils::serialization::DigestBytes;
 
@@ -1064,7 +669,7 @@ mod tests {
             .read_vec(SerKind::Proof, "binding_bytes", binding_len)
             .unwrap();
         let decoded_binding =
-            super::deserialize_composition_binding(&binding_bytes).expect("decode binding");
+            CompositionBinding::deserialize_bytes(&binding_bytes).expect("decode binding");
         assert_eq!(decoded_binding, proof.composition().clone());
         let openings_len =
             read_u32(&mut cursor, SerKind::Openings, "descriptor_len").unwrap() as usize;
@@ -1088,17 +693,20 @@ mod tests {
             .read_vec(SerKind::Openings, "openings_bytes", openings_len)
             .unwrap();
         let decoded_descriptor =
-            super::deserialize_openings_descriptor(&openings_bytes).expect("decode openings");
+            OpeningsDescriptor::deserialize_bytes(&openings_bytes).expect("decode openings");
         assert_eq!(decoded_descriptor, proof.openings().clone());
         let fri_bytes = cursor.read_vec(SerKind::Fri, "fri_bytes", fri_len).unwrap();
-        let decoded_fri = super::deserialize_fri_handle(&fri_bytes).expect("decode fri");
+        let decoded_fri = FriHandle::deserialize_bytes(&fri_bytes).expect("decode fri");
         assert_eq!(decoded_fri, proof.fri().clone());
         if has_telemetry == 1 {
             let telemetry_bytes = cursor
                 .read_vec(SerKind::Telemetry, "telemetry_bytes", telemetry_len)
                 .unwrap();
-            let expected =
-                serialize_telemetry_frame(proof.telemetry().frame()).expect("serialize telemetry");
+            let expected = proof
+                .telemetry()
+                .serialize_bytes()
+                .expect("serialize telemetry")
+                .expect("telemetry bytes");
             assert_eq!(telemetry_len, expected.len());
             assert_eq!(telemetry_bytes, expected);
         }
